requirements.txt
openai
llama-index
supabase
supabase
requests
python-dotenv
streamlit
python-telegram-bot